Add basic schema file
This is just a prototype schema for use with PostgreSQL for setting up
Jim. This is far from complete and there are some non-standard-SQL
things that are incompatible with sqlite3. This will need to be changed
and any bot code will need to be updated to make it more uniform.